revise python tests and example
